Rename component and add examples
